Fix incorrect dashboard_url credential check
The change to introduce required credential checks introduced a bug in
the Spring Insight Framework.  In that case, the check was for
'dashboard-url', which was incorrect.  The correct key to check for
was 'dashboard_url' (note the underscore instead of the hypen).  This
change updates the key to look for and the tests associated with it.

[#64765506]