Clarify non-config component docs
Previously there were a couple of components that did not have
configuration files by default.  Some of these components _could_ be
configured, but were not by default.  The documentation for these
components indicated that configuration should only be modified.  This
change updates the documentation to mention that the configuration
files may need to be created as well as potentially being updated.

[#63758260]