Add Grails documentation
Previously there was no mention of Grails applications anywhere in the
documentation.  This was because Grails applications are treated as
Servlet applications and therefore didn't require anything extra.
Grails users were confused so it needed to be improved.  This change
adds Grails to the first bit of body text in the README and adds an
example of how to run Grails applications.

[#63612598]