Improve Tomcat logging format and location
Previously, the Tomcat logs (and that of anything else that used JULI)
were written in an unreadable form to stderr.  This made the logs
difficult to use when diagnosing problems, but also very scary looking
when the started with [ERR].  This change updates the buildpack to use
a new copy of the Tomcat support JARs which, when combined with new
logging configuration, will cause the output to be cleaner and be
written to stdout.

[#64764934]