Additional JARs on classpath
As part of the work to integrate Ratpack into the buildpack, an
adjustment to the Groovy classpath has been made.  Previously, Groovy
applications were responsible for maintaining their own classpath.
Instead, the buildpack should be doing this.  The new behavior is that
all JAR files (regardless of their location) will be added to the
Groovy classpath at runtime.

[#63888002]