Add http_proxy support to the download_cache
This is for the benefit of Ruby 1.9.3; in Ruby 2.0 Net::HTTP will
automatically use any proxy defined by $http_proxy.

This change also makes sure that request objects only contain the
request part of the uri and not the hostname and port number to
avoid breakage in the proxy use case.