start_script module function
Previously, the start_script method was defined in a module for
inclusion.  Due to my lack of understanding about module methods in
Ruby, I did not realize that this mean that the method could not be
called directly.  This change makes the start_script method a
module_function allowing it to be called directly.