Bash expressions do not fail 'single entry' check
Previously, if the value of a JAVA_OPTS was a bash expression (like in
the AppDynamics Framework), it would run afoul of the enforcement that
values in JAVA_OPTS must be a 'single entry'.  It would, however, only
cause a problem if a custom JAVA_OPT was specified as well (via the
java_opts.yml file). This change updates the detection logic for the
'single entry' so that it makes an exception for bash expressions and
ensures that the check is only run when Play applications are being
pushed.

[#67900270]