Non-webapp Spring Insight Analyzer
Previously Spring Insight required a separate web application (the
analyzer), installed as a peer to a user's primary application, in
order to function.  This meant that Spring Insight didn't conform to
the Cloud Foundry expectation that only a single application run in a
container.

This change integrates the next generation of the Spring Insight
analyzer which is implemented as a Java Agent.  It removes the
addition of the analyzer web application, and also updates the weaver
from a generic vFabric weaver to a Cloud Foundry specific weaver.

This changes leaves outstanding the possibility that it will be used
with Tomcat 8.  In the case that a Spring Insight service is bound to
an application running in Tomcat 8, it will cause the application to
fail during startup.  This issue will be addressed in a later commit.

[resolves #42]