Download latest dependencies
This is the third step in an effort to create a way to build offline-
only buildpacks.  This change uses the buildpack's DownloadCache and
VersionResolver to populate a cache in the staging area.  This cache
contains the latest version (as identified by the configuration) of
all of the dependencies that might be used by the buildpack at
runtime.

As this logic was so complex, this and the other offline build tasks
were broken out into their own Task classes.

[#67851722]