Indicate that a file was retrieved from the cache
Previously, when a file was retrieved from the cache, it displayed a
very short download time (most likely 0.0s).  While this was true and
properly hid the fact that caching was happening, some users were
freaked out by the lack of knowledge about the cache.

This change indicates (using a 'found in cache' message) that a cached
file has been used.  If the file is downloaded it continues to display
the download time as it always has.

[#67662796]