Package task
This change adds a `package` rake task to the buildpack.  This task
creates a .tar.gz of the buildpack, suitable to use with `cf create-
buildpack` and `cf update-buildpack`.  The task is configurable
allowing the user to change the version string embedded in the name of
the output file.  It can also be configured to create an offline
buildpack which embeds the latest version of its dependencies and
disables remote downloads.

[#67852640]