Embedded Version
Previously, it was quite difficult to determine what version of the
buildpack was being used.  In the case that buildpack was cloned, it
would report the remote and hash of the clone, but when the buildpack
was used from the system (i.e. the submodule in dea_ng), there was no
information.

With the push towards a packaged buildpack, the possibilities for
versioning were opened up.  The package now creates a
config/version.yml file which contains the original git hash, git
remote, whether it was created as an offline buildpack, and the
version the user specified on creation.  If this file exists, the
buildpack returns this information in the form:

```
2.1.2 (offline) | https://github.com/cloudfoundry/java-buildpack.git#12345
```

The buildpack does retain the ability to report some version
information when it is not packaged.  In the case that a clone is
used, the version information is reported in the form:

```
https://github.com/cloudfoundry/java-buildpack.git#12345
```

Finally, if somehow the buildpack is run not packaged and not cloned,
the version information is reported as:

```
unknown
```

In short order, the buildpack will be removed from dea_ng and will
only be added as a package.  In this case, the display of the unknown
form will be replaced by the packaged form.  It's likely that users
living on the bleeding edge will continue to see the clone form.

[#68410854]