Do not sort JAVA_OPTS
Previously, the gremlins in my head drove me to sort the JAVA_OPTS
before they were used during the startup of an application.  As
indicated in #41, not only wasn't this needed, it actually prevented
certain use-cases.  This change removes the sorting of the JAVA_OPTS
and updates the associated tests (the much larger bit of work).

[#67684882][resolves #41]