AppDynamics custom resources
Previously, if a user wanted to override the app-agent-config.xml file
in their application, it was impossible to do so.  This change adds
the copy_resources() functionality to the AppDynamics framework.  This
functionality now allows a user to add files to
resources/app_dynamics_agent and have those files overlayed on the
framework's sandbox.

[#67769548]