Document uses of resources directory
Previously there were a number of components that took advantage of
the copy_resources() functionality for configuration.  This facility
is intended to be used by people who fork the buildpack as well, but
wasn't documented anywhere.

This change updates the documentation to indicate where the resources
directory can be used and how to use it.  The documentation also
includes some examples of things you use the resources directory for
(e.g. JCE Unlimited Security).

[#68345350]