Do not log errors before log initialization
There has been a long-running bug in the buildpack caused when
initialization of the Application failed, the failure was swallowed.
The initial failure was swallowed because, as the buildpack attempted
to log the failure, it ran into a situation where the logger had not
yet been initialized.

This change handles that possibility by simply not logging the failure
when the logger has not been initialized yet.  In practice this
failure only happened locally (never on Cloud Foundry) and was driven
by directory permissions problems.