JAVA_OPTS from environment variables
Previously, the JAVA_OPTS environment variable was ignored in all
cases.  This change adds support for integrating the contents of the
JAVA_OPTS environment variable when configured to do so.  By default,
the contents will be used, but a change to the config/java_opts.yml
file by a buildpack admin can disable this behavior.

[#64162734]