Configurations without nested hashes
With the addition of the `config/java_opts.yml` the offline package
build was broken.  The root cause is that when the packager detects a
configuration without `version` and `repository_root` it then attempts
to inspect the values of the configuration as if they
were nested hashes (like the `config/tomcat.yml` file).  The addition
of the `java_opts` configurations was the first time we've had a
configuration that didn't meet the typical conventions and the
packaging build broke.

This change inspects the values of a non-repository-configuration and
only inspects the values if they are also hashes.

[#69170446]