Improve clean semantics
The difference between clean and clobber in rake is quite subtle and
difficult for most users to grok.  This change makes all created
artifacts (regardless of whether are a scratch artifacts or
deliverables) part of CLEAN.  This means that a user will be able to
run either `rake clean` or `rake clobber` and get the same effect.

[#69170656]