Disable Auto-reconfiguration via configuration
Previously, auto-reconfiguration was always enabled.  The auto-
reconfiguration log might sometimes choose not to reconfigure, but if
there was a bug or some other situation that would prevent an
application from running, there was no way to explicitly disable auto-
reconfiguration from even being downloaded and run.

This change adds that configuration option to all three kinds of auto-
reconfiguration and documents how to disable them.

[#72265294]