Package with Ruby 2.1
Previously, the Gemfile of the project assumed that all work was being
done against Ruby 1.9.x.  The CI tests ran successfully on the 2.1
line because they excluded the version-specific debug gems with a
--without debug.  The instructions (and indeed good user experience)
did not ask users to do this, so when bundle install was run in a
non-1.9 Ruby runtime, the gems would not install.

This change updates the Gemfile to install version-specific debug gems
allowing development against 1.9.x, 2.0.x, and 2.1.x.