Empty classpath improvements
Previously, when there were no additional libraries, the classpath
would render as '-cp '.  This was a problem when that argument was
followed by something else.  Applications would end up with something
like '-cp  app.groovy' which was incorrect.  This change updates the
as_classpath() method to return nil if there is are no additional
libraries ensuring that the '-cp' argument is never added to the
command.