Session replication should accept host as well as hostname
Redis provides host and not hostname as a credential for session replication.
The buildpack currently only accepts hostname. This commit adds support for
either 'host' or 'hostname' but not both to be provided in the credentials.

See: https://groups.google.com/a/cloudfoundry.org/d/msg/vcap-dev/vIHzLxZHQ2A/YyT-a6va-cYJ

[#73611494]