Support configurable Tomcat Access Logging
Currently, in order to enable access logging the buildpack must be forked and
a custom access valve added to the Tomcat configuration in the resources
folder. A customised Access Valve is required as the one provided with Tomcat
does not support logging to stdout which is required for Loggregator. This
change allows Access logging to be enabled with a simple configuration change
to the Tomcat container configuration in the 'config' folder. The default
pattern for the logging output is specified in the server.xml file located in
the 'resources/tomcat' folder. The default pattern begins with '[ACCESS]' and
includes the 'X-Vcap-Request_Id' request header but otherwise follows the
default for Apache access logs.

[#72881930]