Support New Relic on Java 8
New Relic has experimental support for Java 8 that isn't included in the
buildpack. This change adds support for New Relic on Java 8 by adding the
appropriate flag to JAVA_OPTS when Java 8 is selected and the application is
bound to a New Relic service.

See: https://discuss.newrelic.com/t/a-note-on-java-8-support/1199/2

[#75561442]