AppDynamics Tier Name in Credentials
Previously, the tier name provided to the AppDynamics agent was hard-
coded in the component configuration file.  This meant that the only
way to change the tier name was to fork the buildpack; a heavy-weight
exercise.  This change makes the configured tier name, the "default"
choice which can be overridden with a 'tier-name' entry in the service
credentials.

[#75252498][#55][#59]