Override App Dynamics tierName from credentials
The tierName provided to App Dynamics is hard coded in the config yml file and
it is beter to support an optional override with a new property in the
credentials of the bound App Dynamics service. This commit changes the
configuration property to 'default_tire_name' and allows it to be over-ridded
with a 'tier_name' property in the App Dynamics service credentials.

See: https://github.com/cloudfoundry/java-buildpack/issues/55
See: https://github.com/cloudfoundry/java-buildpack/issues/59

[#75252498]