HTTPS Repository
Previously, the default repository root pointed at the HTTP endpoint
of download.run.pivotal.io.  This was due to the fact that the HTTPS
certificate (provided by CloudFront) did not match the URI.  The
CloudFront distribution has now been updated with a matching HTTPS
certificate, so this change updates the root to point at the HTTPS
endpoint.

[#77446526]