Conditional Tomcat Configuration
In an earlier, naive, attempt to make Tomcat 8 the default, the code did not take into account that certain
configuration options have changed.  This commit fixes the code so that resource linking is done properly depending on
whether Tomcat 8 or pre-Tomcat 8 is used and only adds the JasperListener to pre-Tomcat 8 is used.

[#78317410][resolves #95]