Output killjava.sh to console
Previously, the killjava.sh script output was directed to a file.  This
worked previously with the crashlogs command to ensure that users could
see what had gone wrong.  With the removal of the crashlogs command,
this command no longer worked properly.  This change updates the script
to output to the console instead, where the output will be picked up by
loggregator.

[#57139852][resolves #79]