Custom Download CA Certificates
Previously, when a download was done using the DownloadCache using
HTTPS, it was assumed that the SSL certificate presented by the server
could be validated with the CA store packaged with OpenSSL.  This
didn't work for servers that included self-signed, or non-standard CA
-based certificates.  This change updates the DownloadCache to allow a
user to present a collection of additional CA certificates.  By
creating a /resources/ca_certs directory and placing the certificates
(in PEM format) within it, the DownloadCache will use the provided
certificates during verification.