Use ca_file instead of ca_path
The previous implementation of CA certs during download used the
ca_path HTTP option.  It appears that this does not work with the
latest versions of OpenSSL, and so it was changed to  use ca_file
instead.  This change also means that all certificates should be
placed in a /resources/ca_certs.pem file instead of the previous
directory strategy.