Tomcat Redis Session Replication
When the Tomcat Redis sub-component detects a service with a name, label, or
tag that has `session-replication` as a substring but both 'host' and
'hostname' as credentials it will fail to use the service. This commit changes
the service detection logic to require at least one credential from a group
instead of exactly one credential from a group. It is up to the individual
containers to handle the credentials provided. In the case of the Tomcat Redis
sub-component 'hostname' will win but this is not documented and users should
not rely on that behaviour in the instances where both 'host' and 'hostname'
are provided with different values.

[#80362118]