Add support for JRebel
This commit adds support for using JRebel in an IDE to live edit an
application running in Cloud Foundry. No service is required, only
the presense of the rebel.xml and rebel-remote.xml files in the
application that get's uploaded to Cloud Foundry. These files are
often generated by the IDE's JRebel plugin.

[#84050588]