Reduce log message for container supports problems
When a container checks its support for a given application its
filter may match a bound service due to the user provided name but
then subsequently fail when the required credentials are not present.
This can happen legitimately as multiple containers can provide
overlapping functionality and require similar filters to identify
support.

[#88312476][#142]