Safenet Luna Security Provider
The change adds a framework for the Safenet Luna Security Provider.
This  framework is triggered by a service with the name 'luna' in it.
It depends on that service having the HSM host, host-certificate,
client, and client-certificate.  In also expects users to provide a
repository with the Luna Client binaries in it.

[#96530962][#99962452]