Support nil configuration values
This commit makes a number of changes all required to support `nil` or empty values in the configuration files. Any `key` with a `nil` value will be ignored but as it is present it can be modified with a new value via an environment variable.

Aditional changes ensure offline buildpacks still contain `nil` configuration values after version pinning. Updates the Java Opts suppport to have a `nil` `java_opts` configuration. Add in some additional `nil` configuration for the memory calculator so these can also be updated via environment variables.

closes #202 and closes #204
[#98822486]