Tomcat Context Path
This change updates the Tomcat container to allow a user to configure
the context path that an application is exposed at.  Previously, the
application was always exposed at the root ('/') but this didn't work
for many applications that had hard-coded context-path assumptions. In
addition, changes to the Cloud Foundry router mean that applications
can now be routed via both a host and a context path.

[#101289516]