Better escaping of Java options
This commit makes it possible to use both escaped and unescaped $ and \ characters in the value of a Java option. The $ character is required so users can include environment variables in Java options. The \ character is required so it's possible to include an escaped $ charecter in the value of a Java option.

[#100119572]