Correct VCAP_ Parsing
Previously the contents of the VCAP_APPLICATION and VCAP_SERVICES
environment variables were parsed as if they were YAML.  It would
appear that YAML flow syntax is an extremely close approximation of
JSON and therefore the parser worked properly.  It was recently
discovered on a new system that there are cases where YAML flow and
JSON diverge causing the parsing to fail.  This change fixes the
parsing to use JSON as it should have.