Add a simple example of adding command line arguments
Command line arguments are quite useful and some apps will not start without them, so this seems like a useful example. The explanation of why it needs to be an array (not a hash) is still really unclear, but at least with this example people will know what to do.