Update environment variable config style
It is possible to specify environment variables that update the
buildpack configuration using both [...] and {...} styles, the
buildpack will handle both. Previously the documentation recommended
the [...] approach. This commit changes the documentation and tests
to use the {...} style. No changes have been made to the
implementation and both [...] and {...} styles are still valid.

[close #237]