Add support for JVMKill agent in Java Buildpack
Provide the ability to use JVMKill agent
(https://github.com/airlift/jvmkill) instead of the regular shell script
that was invoked as an OnOutOfMemoryError handler. This change aims to
make possible for the JVM to recover when the application spawns too
much threads (circumstance that prevents the OnOutOfMemoryError handler
from running).

Issue: #161