Exec commands
A recent change to the execution commands had an unintended side-
effect.  By placing the memory calculator command in a separate
process, connected via an &&, the process tree within a container
changed.  This change was mostly harmless except for the fact that
only the root process (which was no longer the container process) is
signaled for termination.

This change updates all of the container release commands to prepend
themselves with the bash command exec.  This should cause the process
tree to match its earlier behavior of making the container the root
process.