Handle Encoded Downloads
Previously, all downloads passed through DownloadCache required that the value
of the Content-Length header matched the size of the file that was actually
downloaded. This did not take into account compressed encoding schemes for the
artifacts and would fail when they were compressed.  This change detects when
an artifact is compressed and skips the  size validation.

[#116036875]