Deploy Complete Ruxit Agent ZIP
Previously, the agent zip that was downloaded by the Ruxit framework contained
more content than was required by running applications.  Because of this, it
was filtered to remove unnecessary content.  Recent versions of the agent ZIP
file now contain only what is required by the application and therefore this
filtering is unneeded.  By removing this filtering it also makes it possible
to change the required contents of the agent without requiring a change to the
buildpack.

[resolves #286]