End With Wildcards
Previously, the version wildcarding system only worked as an all or nothing
proposition.  This meant that you couldn't ask for a wildcarded suffix to a
version qualifier.  This hadn't been a problem up until now, but with the
latest release of the JRE (_102) the lexical analyzation would not select the
proper version because it came before _91.  This change updates the wildcard
behavior so that it can be used *as a tailing wildcard* allowing us to set a
version of `1.8.0_10+` and select these new versions of the JRE.

[resolves #324]