AppDynamics documentation update
Previously the AppDynamics documentation listed most of the credential
parameters as Optional.  While this is strictly true when viewed across all
supported versions of AppDynamics, for given versions (4.2 for example), some
of those parameters are now required.  This change adds a note to the
documentation indicating that optionality is governed by the version of agent
being used and links to the AppDynamics Java Agent Configuration documentation
for further details.

[resolves #339]