Fail When Unable to Cache Files
Previously, when a file could not be cached during offline packaging, a
warning was printed, but the packaging continued without a necessary file.
This is by design for an older requirement that would allow buildpacks to be
created even in the face of networks that prevented access to certain hosts.
In practice, this requirement hasn't been necessary (disabling components is
the preferred way to deal with this) and it is more dangerous to allow a
package build to complete without all of the necessary files.  This change
updates the warning message to be a hard failure of the packaging.