Dynatrace can always download
Recently, the Dynatrace framework was changed to download, not from a
repository but from the Dynatrace service itself.  This means that even in
completely offline scenarios, its dependencies are still downloadable.  This
change updates the code to remove some unnecessary repository declarations and
updates the code with a guard that ensures that it is always treated as
online.

[resolves #359]