Update release alias
Previously, the release alias rendered the *exact* command returned by the
buildpack.  However, this command now includes an `exec` call which means that
when the process ends, the terminal it is run in, is terminated.  This is
correct behavior for running within a container, but undesirable when running
locally.  This change excludes removes the exec from the command as part of
the release alias.