Updated Container Certificate Trust Store
Previously the Container Certificate Trust Store framework did a bunch of
parsing work in Ruby and then called the keytool application repeatedly for
each of the ~170 certificates in a container.  This led to it being incredibly
slow, on the order of 1m to process all certificates.  Because of this it was
disabled by default.  This change adds a full-Java parser and KeyStore creator
to the mix and decreases the amount of time required to create the KeyStore to
~0.3s.  Because of this it is now enabled by default.