Fix wildcard resolving for version qualifier
Added a new criteria for qualifier comparison. If the qualifier has a number at the beginning of it's string, that number is used for comparing. Otherwise, if there is no numeric symbols in the beginning of qualifier's string, or those symbols represent equal numbers, then the old criteria applies, meaning that the qualifiers are compared as strings.

Issue: #397