Make JRE Source Optional
Previously, the JRE source was both required and fixed to a location only
available in the "normal" JRE.  This caused an issue where the JRE source
could not be found in its usual place (because it was a server JRE).  This
missing JRE source caused a staging failure.  This change updates the logic to
try the "normal" JRE location, then the server JRE location, and if neither
are found to simply not add any JRE source at all, allowing staging to
complete without integrating the JRE certificates.

[resolves #415]