Remove Darwin Binaries from Offline Buildpack
Previously, Darwin binaries were included in the builds of offline buildpacks.
While it's important for development and testing that these binaries exist,
they do not need to be packaged into the offline buildpack that is distributed
for Cloud Foundry.  This change removes them, saving some space in the offline
buildpack.