Remove Most UTF-8 Characters
Previously, there were a handful of UTF-8 characters (and one specifically in
an error message) in the repository.  This caused issues when the removal of
explicit Ruby encodings happened as some other parties were still testing
against Ruby 1.9.  This change removes most of the UTF-8 characters in the
repository, leaving a single test that is specifically looking at special
UTF-8 characters.