Compile JRE First
Previously, the JRE was the first component to be laid down on disk during the
compile phase which enabled all subsequent components to use it if necessary.
In a recent change, this ordering was modified such that it ran after
frameworks preventing them from using the JRE during compile.  This change
moves the JRE back to its original position and adds additional security
provider functionality that was formerly enabled by the move.

[resolves #432]