Polishing
This change does some polishing to this contribution in order to bring it in
line with current coding standard.  Of note, this removes the jvmkill
integration because the contributions needs to go into the 3.x branch where
that does not exist.  The jvmkill integration will be added back when this
contribution is merged from 3.x to master.

[resolves #427]