Fix agentManager.url.1 property, add support for related older properties, enable introscope agent by default and update default repository_root
Before this commit, the Introscope agent framework had support for the new agent manager url, but it was missing a few cases.
The framework was also disabled by default, so changes had to be made to enable it.
The config/introscope_agent.yml file had an empty repository_root.

The biggest change was for the agentManager.url.1 property. Since it is now just a url, separate properties for host,
port, and ssl do not need to be maintained. All information is encapsulated in the url. The user provides the url, and
to allow backwards-compatability, the previous EM connection properties will also be populated with the various components
from the agentManager url.

After this commit, the buildpack simplifies the user-input so that the user only needs to provide the url, and
not the host-name and port. This way, the detection criteria has also been changed so that the url must be provided,
and not the host-name. The previous versions of the buildpack were missing support for different socket factories,
and with this commit, all socket factories and connection methods (tcp, ssl, http, https) are now supported. Finally,
the introscope agent has been enabled out of the box, and the repository root now contains a link to the index.yml file.
The documentation has also been updated to reflect these changes.