Cached Filename of Constant Length
Previously, the name of a cached file was a sanitized version of the URI.
Recently, a user starting caching a file that had a URI that was longer than
255 characters, causing issues with writing a filename of that length to the
filesystem.  The solution to this was to hash the URI (SHA256 in this case) in
order to get a shorter and constant-length filename.

[resolves #484]