One Service With Credentials
Previously, the Services.one_service? method ensured first that there was a
single service that matched and then ensured that the credentials of that
service matched.  This was more restrictive than it needed to be, so this
change loosens things a bit such that it now enforces only a single service
that matches name, label, tags, and required credentials.

[resolves #489]