Large Output Shell Stall
Previously if a large amount of output was received during a shell call, the
buildpack would stall because the process' output buffer would fill to
capacity before the process would complete.  This change assigns the output to
a variable during execution, ensuring that the process' output buffers do not
fill during execution.

[resolves #507]