Reduce Expected Thread Count
Previously, the expected thread count (used for memory calculations) was set
at 300.  This was the Tomcat default 200 plus 100 "other" threads.  I've been
suitably convinced that the "other" collection is closer to 50 than it is to
100 even accounting for connection pools and the like.  This change updates
the default to 250 which will reclaim 50M be default, and we'll see how it
goes.