Container Security Provider Manager Flags
Previously, the Container Security Provider could only be enabled and disabled
completely for a given application.  In certain circumstances it was useful to
disable the KeyManager and TrustManager individually.  This change takes
advantage of a new feature in the Container Security Provider itself and sets
the appropriate flags to disable the managers independently.

[resolves #552]