Support NewRelic Extensions
This commit adds support for Custom NewRelic Instrumentation via
extension files (typically XML/JAR files)

The desired extensions are to be placed in a directory and tarred up.
The tarfile can then be hosted in a repository in the usual way (with
`repository_root` and `version` specified in a sub-configuration called
`extensions`).

The buildpack will download this tarfile if `extensions.repository_root`
is configured, and will unpack the files into the 'extensions' directory
where the NewRelic agent will look for them by default.