Generic Introscope Properties
Previously in order to handle Introscope properties, a new value would have to
be added to the framework for integration.  This made arbitrary configuration
difficult and brittle.  This change updates the framework to behave like many
of the others and do a generic mapping from any unknown credential payload
property to a system property with a fixed prefix.

[resolves $717]

Signed-off-by: Ben Hale <bhale@pivotal.io>