Riverbed Always Connects to Service Broker
Previously when the Riverbed integration was trying to download an agent from
the service broker, it could be stopped by an offline buildpack.  Given that
download from the service broker should always work (offline is relative to
the internet, not the internal network), this change enables the framework to
download from the service broker no matter what.

[resolves #734]

Signed-off-by: Ben Hale <bhale@pivotal.io>