Remove Redis Store and Add Redisson
The Tomcat container has support for using Redis to persist sessions. It was previously using the Redis Store project for this, however, that project has been end-of-life for a long time. As such, we've removed the usage of Redis Store and swapped in the Redission library instead. It also provides support for Tomcat session storage in Redis. This project is currently maintained and provides a better outlook for users that need to run Tomcat and store sessions in Redis.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>