Remove Takipi Agent (#956)
The download site for the Takipi Agent has gone dark, and attempts to alert the vendor and rectify the situation have failed. This component will no longer be enabled with the default distribution of the Java buildpack. You can re-enable it if you are able to host your own repository for the downloads.

If/when the vendor is able to resolve the situation, we can re-enable the agent by default in the buildpack.

Signed-off-by: Daniel Mikusa <dmikusa@vmware.com>